Добавила класс AdsControllerTest, и пока нерабочий тест на addAds()

+++ src/main/java/ru/skypro/homework/dto/AdsDto.java
@@ -12,7 +12,7 @@ public class AdsDto {
 
     private String image;
 
-    private int pk;
+    private long pk;
 
     private int price;
